refactor: decompose App component
